add: new files are added by updating a create project script.

+++ create_project.py
@@ -25,11 +25,11 @@ files = [
     f"{project_name}/src/loaders/csv_loader.py",
     f"{project_name}/src/loaders/json_loader.py",
     f"{project_name}/src/loaders/sql_loader.py",
-     f"{project_name}/analysis/analyze_src/basic_data_inspection.py",
-    f"{project_name}/analysis/analyze_src/missing_values_analysis.py",
-    f"{project_name}/analysis/analyze_src/univariate_analysis.py",
-    f"{project_name}/analysis/analyze_src/bivariate_analysis.py",
-    f"{project_name}/analysis/analyze_src/multivariate_analysis.py",
+     f"{project_name}/src/analysis_src/basic_data_inspection.py",
+    f"{project_name}/src/analysis_src/missing_values_analysis.py",
+    f"{project_name}/src/analysis_src/univariate_analysis.py",
+    f"{project_name}/src/analysis_src/bivariate_analysis.py",
+    f"{project_name}/src/analysis_src/multivariate_analysis.py",
     f"{project_name}/tests/test_csv_loader.py",
     f"{project_name}/tests/test_json_loader.py",
     f"{project_name}/tests/test_sql_loader.py",
@@ -43,7 +43,8 @@ for folder in folders:
 
 # Create empty files
 for file in files:
-    with open(file, 'w') as f:
-        pass
+    if not os.path.exists(file):
+        with open(file, 'w') as f:
+            pass
 
 print(f"Project '{project_name}' structure created successfully!")
